requirements.txt
fastapi
uvicorn
python-dotenv
chardet
requests